Ignore string prefixes like "b" or "u"
When the string uses prefixes, the checker wasn't checking those strings,
because it only handles strings that start with a bad quote. To handle this,
it gets the last character of the string (which is the quote character used)
and removes all characters in front of the first quote character.

It also adds a few tests using one or two prefixes.