Repaired stdin support by moving from `pep8` to `flake8` (#38)
* Require `flake8` and use its `pep8` import

Instead of relying on `pep8` importing itself, it can just use Flake8's
`pep8` import, which will either map to `pep8` itself or to `pycodestyle`
depending on what Flake8 is using. This also requires Flake8 to be installed.

* Test using stdin as input

Using stdin as input was broken with version 2.6 of Flake8. This test should
prevent it happening again for future versions of the plugin.

* Require non-multithreading for Flake8 3.x test

Because of a current issue in Flake8 3.x (no 164), it causes a `PicklingError`
when requiring `flake8` and running the tests. To avoid this it is setting the
number of jobs to 1 which will avoid any that the code causing the error is
executed.

* Test using setup.cfg instead

* Restore original .travis.yml

As all the changes from `.travis.yml` have been removed, the original file can
be used again.