Use flake8-polyfill to support Flake8 2.x and 3.x
As Flake8 2.6 switched from `pep8` to `pycodestyle` so that using `pep8`
itself doesn't help. So with 09198d1e the plugin imported `flake8.engine.pep8`
which is either one of them depending on the version. But Flake8 3.x doesn't
have this package anymore so it broke support for Flake8 3.x then.

This is redesigning the import to use `flake8-polyfill` which provides an
easily handleable version number. So instead of using `flake8.engine.pep8` it
is comparing the version number and based on that either imports `pep8` or
`pycodestyle`. But additionally to that it needs to monkey patch the stdin
function in Flake8 3.x into that package.

It enables the test using Flake8's stdin functionality on Flake8 3.x. It also
adds a test to use Flake8 on a file.