Use proper entry point name
The entry point must be of the form `errorcode = target` so that it is
possible to select the plugin explicitly in Flake8 2.x and implicitly in
Flake8 3.x.

The following commands would've yielded no errors from this plugin without the
change. In Flake8 2.x it wasn't possible to select it:

    flake8 --select=Q FILES

In Flake8 3.x (if the code has been made compatible again) it wasn't selected
by default without the user explicitly selecting it:

    flake8 FILES

Calling the command in the other Flake8 version would have worked.